requirements.txt
requests
python-dotenv
nhl-api-py
python-telegram-bot
